micro debug

+++ src/main/java/org/application/lab5/dragons/DragonHead.java
@@ -31,6 +31,6 @@ public class DragonHead {
      */
     @Override
     public String toString() {
-        return "Has " + eyesCount + " eyes";
+        return "Has " + eyesCount.intValue() + " eyes";
     }
 }
